Added token check for users

+++ main.py
@@ -6,10 +6,23 @@ import hashlib
 
 app = Flask(__name__)
 
+def check(token):
+	conn = mysql.connect()
+	cursor = conn.cursor()
+	query = """SELECT * FROM Token WHERE Hash=%s"""
+	cursor.execute(query,(token))
+	conn.commit()
+	data = cursor.fetchall()
+	if len(data):
+		return 1
+	else:
+		return 0
+
 
 @app.route("/")
 def login():
 	return render_template('index.html')
+
 
 @app.route("/login",methods=['POST'])
 def token():
@@ -31,7 +44,6 @@ def token():
 				return "NOT Added"
 			else:
 				return render_template('token.html',has=_hash)
-
 	except Exception as e:
 		return json.dumps({'error':str(e)})
 	finally:
@@ -43,20 +55,25 @@ def token():
 def add():
 
 	try:
+		token = request.args.get('token') 
 		name = request.args.get('name')
 		price = request.args.get('price')
 		flag = request.args.get('flag')
 		conn = mysql.connect()
 		cursor = conn.cursor()
-		if name and price:
-			query = """INSERT INTO User (Name, Price, Available) VALUES (%s, %s, %s)"""
-			cursor.execute(query,(name,price,flag))
-			conn.commit()
-			data = cursor.fetchall()
-			if data is None:
-				return "NOT Added"
-			else:
-				return "Added"
+		
+		if token and check(token):
+			if name and price:
+				query = """INSERT INTO User (Name, Price, Available) VALUES (%s, %s, %s)"""
+				cursor.execute(query,(name,price,flag))
+				conn.commit()
+				data = cursor.fetchall()
+				if data is None:
+					return "NOT Added"
+				else:
+					return "Added"
+		else:
+			return render_template('unauth.html')
 	except Exception as e:
 		return json.dumps({'error':str(e)})
 	finally:
@@ -68,29 +85,34 @@ def add():
 def delete():
 
 	try:
+		token = request.args.get('token') 
 		name = request.args.get('name')
 		lt = request.args.get('lt')
 		gt = request.args.get('gt')
 		conn = mysql.connect()
 		cursor = conn.cursor()
-		if name:
-			query = """DELETE FROM User WHERE name=%s"""
-			cursor.execute(query,(name))
-			conn.commit()
-			data = cursor.fetchall()
-			if data is None:
-				return '0'
-			else:
-				return '1'
-		elif (lt and gt):
-			query = """DELETE FROM User WHERE (price>=%s and price<=%s)"""
-			cursor.execute(query,(lt,gt))
-			conn.commit()
-			data = cursor.fetchall()
-			if data is None:
-				return '0'
-			else:
-				return '1'
+		
+		if token and check(token):
+			if name:
+				query = """DELETE FROM User WHERE name=%s"""
+				cursor.execute(query,(name))
+				conn.commit()
+				data = cursor.fetchall()
+				if data is None:
+					return '0'
+				else:
+					return '1'
+			elif (lt and gt):
+				query = """DELETE FROM User WHERE (price>=%s and price<=%s)"""
+				cursor.execute(query,(lt,gt))
+				conn.commit()
+				data = cursor.fetchall()
+				if data is None:
+					return '0'
+				else:
+					return '1'
+		else:
+			return render_template('unauth.html')
 	except Exception as e:
 		return json.dumps({'error':str(e)})
 	finally:
@@ -102,31 +124,36 @@ def delete():
 def find():
 
 	try:
+		token = request.args.get('token')
 		name = request.args.get('name')
 		lt = request.args.get('lt')
 		gt = request.args.get('gt')
 		conn = mysql.connect()
 		cursor = conn.cursor()
-		if name:
-			query = """SELECT * FROM User WHERE name=%s"""
-			cursor.execute(query,(name))
-			conn.commit()
-			data = cursor.fetchall()
-			print data
-			if data is None:
-				return '0'
-			else:
-				return '1'
-		elif (lt and gt):
-			query = """SELECT * FROM User WHERE (price>=%s and price<=%s)"""
-			cursor.execute(query,(lt,gt))
-			conn.commit()
-			data = cursor.fetchall()
-			print data
-			if data is None:
-				return '0'
-			else:
-				return '1'
+
+		if token and check(token):
+			if name:
+				query = """SELECT * FROM User WHERE name=%s"""
+				cursor.execute(query,(name))
+				conn.commit()
+				data = cursor.fetchall()
+				print data
+				if data is None:
+					return '0'
+				else:
+					return '1'
+			elif (lt and gt):
+				query = """SELECT * FROM User WHERE (price>=%s and price<=%s)"""
+				cursor.execute(query,(lt,gt))
+				conn.commit()
+				data = cursor.fetchall()
+				print data
+				if data is None:
+					return '0'
+				else:
+					return '1'
+		else:
+			return render_template('unauth.html')
 	except Exception as e:
 		return json.dumps({'error':str(e)})
 	finally:
@@ -138,43 +165,48 @@ def find():
 def edit():
 
 	try:
+		token = request.args.get('token')
 		bid = request.args.get('id')
 		name = request.args.get('name')
 		price = request.args.get('price')
 		flag = request.args.get('flag')
 		conn = mysql.connect()
 		cursor = conn.cursor()
-		if bid:
-			if name:
-				query = """UPDATE User SET name=%s WHERE id=%d"""
-				cursor.execute(query,(name,bid))
-				conn.commit()
-				data = cursor.fetchall()
-				print data
-				if data is None:
-					return '0'
-				else:
-					return '1'
-			elif price:
-				query = """UPDATE User SET price=%s WHERE id=%d"""
-				cursor.execute(query,(price,bid))
-				conn.commit()
-				data = cursor.fetchall()
-				print data
-				if data is None:
-					return '0'
-				else:
-					return '1'
-			elif flag:
-				query = """UPDATE User SET Available=%s WHERE id=%d"""
-				cursor.execute(query,(flag,bid))
-				conn.commit()
-				data = cursor.fetchall()
-				print data
-				if data is None:
-					return '0'
-				else:
-					return '1'
+
+		if token and check(token):
+			if bid:
+				if name:
+					query = """UPDATE User SET name=%s WHERE id=%d"""
+					cursor.execute(query,(name,bid))
+					conn.commit()
+					data = cursor.fetchall()
+					print data
+					if data is None:
+						return '0'
+					else:
+						return '1'
+				elif price:
+					query = """UPDATE User SET price=%s WHERE id=%d"""
+					cursor.execute(query,(price,bid))
+					conn.commit()
+					data = cursor.fetchall()
+					print data
+					if data is None:
+						return '0'
+					else:
+						return '1'
+				elif flag:
+					query = """UPDATE User SET Available=%s WHERE id=%d"""
+					cursor.execute(query,(flag,bid))
+					conn.commit()
+					data = cursor.fetchall()
+					print data
+					if data is None:
+						return '0'
+					else:
+						return '1'
+		else:
+			return render_template('unauth.html')
 	except Exception as e:
 		return json.dumps({'error':str(e)})
 	finally:
